refactor: dissolve the Pages namespace into Core and Http
PageSchema and the page layout enums are composition primitives, so they
move into Core. Page is the routable Inertia request handler, so it moves
into Http alongside the controllers. Nothing about a page is a feature
domain of its own.
