+++ src/Core/Enums/PageContainer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bambamboole\Lattice\Pages\Enums;
+namespace Bambamboole\Lattice\Core\Enums;
 
 enum PageContainer: string
 {
