+++ tests/Feature/LatticePageTest.php
@@ -31,6 +31,7 @@ use Bambamboole\Lattice\Core\Enums\Gap;
 use Bambamboole\Lattice\Core\Enums\HttpMethod;
 use Bambamboole\Lattice\Core\Enums\LucideIcon;
 use Bambamboole\Lattice\Core\Enums\Width;
+use Bambamboole\Lattice\Core\PageSchema;
 use Bambamboole\Lattice\Core\Toasts\Enums\ToastType;
 use Bambamboole\Lattice\Core\Toasts\ToastMessage;
 use Bambamboole\Lattice\Facades\Lattice;
@@ -40,11 +41,10 @@ use Bambamboole\Lattice\Forms\Components\PasswordInput;
 use Bambamboole\Lattice\Forms\FormDefinition;
 use Bambamboole\Lattice\Fragments\Components\Fragment as FragmentComponent;
 use Bambamboole\Lattice\Fragments\FragmentDefinition;
+use Bambamboole\Lattice\Http\Page;
 use Bambamboole\Lattice\LatticeRegistry;
 use Bambamboole\Lattice\Menu\MenuItem;
 use Bambamboole\Lattice\Menu\MenuRegistry;
-use Bambamboole\Lattice\Pages\Page;
-use Bambamboole\Lattice\Pages\PageSchema;
 use Bambamboole\Lattice\Security\ComponentReferenceSigner;
 use Bambamboole\Lattice\Tables\Columns\StackColumn;
 use Bambamboole\Lattice\Tables\Columns\TextColumn;
