+++ tests/Fixtures/Discovery/DiscoveredPanelFragment.php
@@ -6,8 +6,8 @@ namespace Bambamboole\Lattice\Tests\Fixtures\Discovery;
 
 use Bambamboole\Lattice\Attributes\Fragment;
 use Bambamboole\Lattice\Core\Components\Text;
+use Bambamboole\Lattice\Core\PageSchema;
 use Bambamboole\Lattice\Fragments\FragmentDefinition;
-use Bambamboole\Lattice\Pages\PageSchema;
 
 #[Fragment('fixtures.panel')]
 class DiscoveredPanelFragment extends FragmentDefinition
