+++ workbench/app/Pages/WorkbenchTablesPage.php
@@ -11,8 +11,8 @@ use Bambamboole\Lattice\Core\Components\Tab;
 use Bambamboole\Lattice\Core\Components\Tabs;
 use Bambamboole\Lattice\Core\Components\Text;
 use Bambamboole\Lattice\Core\Enums\Gap;
-use Bambamboole\Lattice\Pages\Page;
-use Bambamboole\Lattice\Pages\PageSchema;
+use Bambamboole\Lattice\Core\PageSchema;
+use Bambamboole\Lattice\Http\Page;
 use Bambamboole\Lattice\Tables\Components\Table;
 use Workbench\App\Tables\UsersInfiniteTable;
 use Workbench\App\Tables\UsersNoneTable;
