+++ workbench/app/Providers/TypeScriptTransformerServiceProvider.php
@@ -8,11 +8,11 @@ use Bambamboole\Lattice\Actions\Enums\EffectType;
 use Bambamboole\Lattice\Core\Enums\Align;
 use Bambamboole\Lattice\Core\Enums\Gap;
 use Bambamboole\Lattice\Core\Enums\HttpMethod;
+use Bambamboole\Lattice\Core\Enums\PageContainer;
+use Bambamboole\Lattice\Core\Enums\PageLayout;
 use Bambamboole\Lattice\Core\Enums\Width;
 use Bambamboole\Lattice\Core\Toasts\Enums\ToastType;
 use Bambamboole\Lattice\Forms\Enums\Op;
-use Bambamboole\Lattice\Pages\Enums\PageContainer;
-use Bambamboole\Lattice\Pages\Enums\PageLayout;
 use Bambamboole\Lattice\Tables\Enums\PaginationType;
 use Spatie\LaravelTypeScriptTransformer\TypeScriptTransformerApplicationServiceProvider;
 use Spatie\TypeScriptTransformer\TypeScriptTransformerConfigFactory;
